move origin rdb_test into clickhouse
